feat(ExperimentSet): Experiments are now stored in the base directory
ExperimentSet no longer creates a new directory for each experiment set.
Instead, the experiments are stored in the base directory. This change
makes it such that changes to `config_set.yaml` do not lead to the
creation of a new directory. Additionally, this change was necessary as
the ConfigSet object was unordered, which led to the creation of new directories
when the order of the keys in the Set changed.